requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
discord.py==1.7.2
idna==3.1
idna-ssl==1.1.0
multidict==5.1.0
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.1
pytz==2021.1
scipy==1.5.4
six==1.16.0
typing-extensions==3.10.0.0
xlrd==2.0.1
xlutils==2.0.0
xlutils3==1.0.2
xlwt==1.3.0
yarl==1.6.3
